// template/react-starter-kit/src/.generated/api/model/noticesControllerGetNoticeFeedDirectionItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type NoticesControllerGetNoticeFeedDirectionItem = typeof NoticesControllerGetNoticeFeedDirectionItem[keyof typeof NoticesControllerGetNoticeFeedDirectionItem];


export const NoticesControllerGetNoticeFeedDirectionItem = {
  asc: 'asc',
  desc: 'desc',
} as const;