// template/react-starter-kit/src/.generated/api/model/noticesControllerGetNoticeFeedSortItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type NoticesControllerGetNoticeFeedSortItem = typeof NoticesControllerGetNoticeFeedSortItem[keyof typeof NoticesControllerGetNoticeFeedSortItem];


export const NoticesControllerGetNoticeFeedSortItem = {
  title: 'title',
  priority: 'priority',
  publishedAt: 'publishedAt',
  expiresAt: 'expiresAt',
  createdAt: 'createdAt',
  id: 'id',
} as const;